backend/requirements.txt
flask==2.3.3
python-dotenv==1.0.0
flask-cors==4.0.0
pytesseract==0.3.10
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.0.0
pandas==2.0.3
python-jose==3.3.0
pytest==7.4.0
gunicorn==21.2.0
pdf2image==1.16.3
mistralai==0.4.2
python-magic==0.4.27
reportlab==4.0.4
pytest-cov==4.1.0 
python-magic-bin==0.4.14